chore(db): Remove the scorecard
Let us have it created by the application instead.

The data we need is already there, and we create the scorecard in the
application, or as a view for the application instead.

I need to spend some serious time now, learning PostGraphile, to get the API for
this a little more clear, and stable, so that I can progress.

Signed-off-by: Ole P. Orhagen <ole.petter.orhagen@mimiro.no>